added apis for book and author

+++ Controllers/Book.js
@@ -1,4 +1,6 @@
 import BookService from "../Services/Book.js";
+import sendEmail from "../email/Email.js";
+import Member from "../Models/Member.js";
 
 class BookController {
     async createBook(req, res){
@@ -55,11 +57,14 @@ class BookController {
         try {
             const {id} = req.params;
             const updates = req.body;
-
+            const findById = await BookService.fetchBookById(id);
+            if (findById.isPublished) {
+                return res.status(400).json({ message: "Cannot update a published book" });
+            }
             if (Object.prototype.hasOwnProperty.call(updates, "isPublished")) {
                 delete updates.isPublished;
             }
-
+            
               if(req.file){
                 updates.coverImageUrl = req.file.path;
               }
@@ -67,6 +72,7 @@ class BookController {
               if (!updatedData) {
                 return res.status(404).json({ error: "Book not found" });
               }
+             
         
               res.status(200).json({
                 message: "Book updated successfully",
@@ -82,10 +88,15 @@ class BookController {
     async deleteBookById(req, res){
         try {
             const {id} = req.params;
+            const findBook = await BookService.fetchBookById(id)
+            if (findBook.isPublished) {
+                return res.status(400).json({ message: "Cannot delete a published book" });
+            }
             const book = await BookService.deleteBook(id);
             if(!book){
                 return res.status(404).json("book not found")
             }
+    
             res.status(200).json("book deleted successfully")
         } catch (error) {
             res.status(500).json({error:"failed deleting book",details:error.message})
@@ -105,6 +116,19 @@ class BookController {
 
               const status = updatedBook.isPublished ?"published" :"unpublished";
 
+              if(updatedBook.isPublished){
+                const subscribedMembers = await Member.find({subscribedBooks: id});
+                for(const member of subscribedMembers){
+                    const subject = `New Book Published: ${updatedBook.title}`;
+                    const text = `Dear ${member.name},\n\nThe book "${updatedBook.title}" has just been published. Check it out now!`;
+                    const html = `<p>Dear ${member.name},</p><p>The book "<strong>${updatedBook.title}</strong>" has just been published. Check it out now!</p>`;
+                    try {
+                        await sendEmail(member.email,subject, text, html)
+                    } catch (error) {
+                        console.error(`Failed to send email to ${member.email}:`, emailError.message);
+                    }
+                }
+              }
               res.status(200).json({
                 message: `Book ${status} successfully`,
                 data: updatedBook,
@@ -114,6 +138,49 @@ class BookController {
             res.status(500).json({
                 error: "Failed to toggle publish status",
                 details: error.message,
+              });
+        }
+    }
+
+    async getPublishedBooks(req, res){
+        try {
+            const {page, limit, genre, language} = req.query;
+            const {books, totalBooks} = await BookService.fetchPublishBooks({page, limit, genre, language})
+
+            return res.status(200).json({
+                message: "Published books fetched successfully",
+                data: {
+                    books,
+                    pagination: {
+                        currentPage: parseInt(page) || 1,
+                        totalPages: Math.ceil(totalBooks / (parseInt(limit) || 10)),
+                        totalBooks,
+                    },
+                },
+            });
+        } catch (error) {
+            return res.status(500).json({
+                message: "Failed to fetch published books",
+                error: error.message,
+            });
+        }
+        
+    }
+    async getBookByIdWeb(req, res){
+        try {
+            const {bookId} = req.params;
+            const {language} = req.query;
+
+            const bookDetails = await BookService.getBookById({bookId, language});
+
+            return res.status(200).json({
+                message: "Book details fetched successfully",
+                data: bookDetails,
+              });
+        } catch (error) {
+            return res.status(500).json({
+                message: "Failed to fetch book details",
+                error: error.message,
               });
         }
     }
